azure-function/requirements.txt
azure-functions>=1.11.0
requests>=2.28.0
